Added Cirrus Consolidation script
This Powershell script let's you combine all UTXOs in the wallet you choose into a single destination address. It takes a default set of 900 UTXOs per round and loops through all UTXOs but leaves the UTXOs in the modulo equation of `Total UTXOs %  Number of UTXOs per Coin Set`. As it sorts the UTXOs descending on Amount, the dust is what is left.

If you have less than 900 UTXOs in your wallet, you can lower the Coin Set Size in the script, but do not set it higher than 1000. You may set it to 1 or 2 in order to run a test round. Or use a debugger.

The script will report which transaction it's doing and has a Pause built in per transaction. You may comment it out if you trust the script to do what it's supposed to.
When a transaction is successful, the TransactionID and outputs (destination address and amount in stratoshis) are printed on screen.